[tools] lib traceeveent: Allow for negative numbers in print format
Message-id: <1455542356-29341-191-git-send-email-jolsa@redhat.com>
Patchwork-id: 134560
O-Subject: [PATCH RHEL7.3 BZ1306834 190/237] tools lib traceeveent: Allow for negative numbers in print format
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 1d945012d1381f0232ea2d66e32b06182eedb476
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Thu Aug 27 09:46:01 2015 -0400

description
===========
It was reported that "%-8s" does not parse well when used in the printk
format. The '-' is what is throwing it off. Allow that to be included.

Reporter note:

Example before:

  transhuge-stres-10730 [004]  5897.713989: mm_compaction_finished: node=0
  zone=>-<8s order=-2119871790 ret=

Example after:

  transhuge-stres-4235  [000]   453.149280: mm_compaction_finished: node=0
  zone=ffffffff81815d7a order=9 ret=

(I will send patches to fix the string handling in the tracepoints so
it's on par with in-kernel printing via trace_pipe:)

  transhuge-stres-10921 [007] ...1  6307.140205: mm_compaction_finished: node=0
  zone=Normal   order=9 ret=partial

Signed-off-by: Rafael Aquini <aquini@redhat.com>