[tools] lib traceevent: Fix string handling in heterogeneous arch environments
Message-id: <1455542356-29341-233-git-send-email-jolsa@redhat.com>
Patchwork-id: 134601
O-Subject: [PATCH RHEL7.3 BZ1306834 232/237] tools lib traceevent: Fix string handling in heterogeneous arch environments
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit c2e4b24ff848bb180f9b9cd873a38327cd219ad2
Author: Kapileshwar Singh <kapileshwar.singh@arm.com>
Date: Tue Sep 22 14:22:03 2015 +0100

description
===========
When a trace recorded on a 32-bit device is processed with a 64-bit
binary, the higher 32-bits of the address need to ignored.

The lack of this results in the output of the 64-bit pointer
value to the trace as the 32-bit address lookup fails in find_printk().

Before:

  burn-1778  [003]   548.600305: bputs:   0xc0046db2s: 2cec5c058d98c

After:

  burn-1778  [003]   548.600305: bputs:   0xc0046db2s: RT throttling activated

The problem occurs in PRINT_FIELD when the field is recognized as a
pointer to a string (of the type const char *)

Heterogeneous architectures cases below can arise and should be handled:

* Traces recorded using 32-bit addresses processed on a 64-bit machine
* Traces recorded using 64-bit addresses processed on a 32-bit machine

Signed-off-by: Rafael Aquini <aquini@redhat.com>