[tools] perf machine: Add pointer to sample's environment
Message-id: <1456127829-13546-63-git-send-email-jolsa@redhat.com>
Patchwork-id: 135310
O-Subject: [PATCH BZ1308570 062/260] perf machine: Add pointer to sample's environment
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 4cde998d205894705b534878122631142a3eefe4
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Sep 9 12:25:00 2015 -0300

description
===========
The 'struct machine' represents the machine where the samples were/are
being collected, and we also have a 'struct perf_env' with extra details
about such machine, that we were collecting at 'perf.data' creation time
but we also needed when no perf.data file is being used, such as in
'perf top'.

So, get those structs closer together, as they provide a bigger picture
of the sample's environment.

In 'perf session', when the file argument is NULL, we can assume that
the tool is sampling the running machine, so point machine->env to
the global put in place in previous patches, while set it to the
perf_header.env one when reading from a file.

This paves the way for machine->env to be used in
perf_event__preprocess_sample to populate addr_location.socket.

Signed-off-by: Rafael Aquini <aquini@redhat.com>