[tools] build: Allow setting the feature detection user
Message-id: <1456127829-13546-80-git-send-email-jolsa@redhat.com>
Patchwork-id: 135326
O-Subject: [PATCH BZ1308570 079/260] tools build: Allow setting the feature detection user
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 13e96db61c1c32cd4c8102a95129bb7677cc746d
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Sep 21 12:24:47 2015 -0300

description
===========
We will use the tools/build/ autodetection in the eBPF patchkit
and it is currently sharing the output directory with perf, that
also uses the feature detection logic.

As we keep state in the output directory, so that we can avoid running
all the tests again, we need to have different filenames for the files
used in this state, allow doing that via the FEATURE_USER variable, to
be set alongside the existing FEATURE_{TEST,DISPLAY} variables.

v2: Fix comment describing the FEATURE_DUMP filename to make sure where
    it is created, precisely at $(OUTPUT)FEATURE-DUMP$(FEATURE_USER).
    Pointed out by Jiri.

Signed-off-by: Rafael Aquini <aquini@redhat.com>