[tools] perf report: Make max_stack value allow for synthesized callchains
Message-id: <1456127829-13546-108-git-send-email-jolsa@redhat.com>
Patchwork-id: 135345
O-Subject: [PATCH BZ1308570 107/260] perf report: Make max_stack value allow for synthesized callchains
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 188bb5e2ce112463428994f91291e5df6fc05521
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Sep 25 16:15:46 2015 +0300

description
===========
perf report has an option (--max-stack) to set the maximum stack depth
when processing callchains.  The option defaults to the hard-coded
maximum definition PERF_MAX_STACK_DEPTH which is 127.  The intention of
the option is to allow the user to reduce the processing time by
reducing the amount of the callchain that is processed.

It is also possible, when processing instruction traces, to synthesize
callchains.  Synthesized callchains do not have the kernel size
limitation and are whatever size the user requests, although validation
presently prevents the user requested a value greater that 1024.  The
default value is 16.

To allow for synthesized callchains, make the max_stack value at least
the same size as the synthesized callchain size.

Signed-off-by: Rafael Aquini <aquini@redhat.com>