[tools] perf inject: Add --strip option to strip out non-synthesized events
Message-id: <1456127829-13546-116-git-send-email-jolsa@redhat.com>
Patchwork-id: 135349
O-Subject: [PATCH BZ1308570 115/260] perf inject: Add --strip option to strip out non-synthesized events
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit f56fb9864c501dc85ebe40af5bf925dd07d990c0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Sep 25 16:15:55 2015 +0300

description
===========
Add a new option --strip which is used with --itrace to strip out
non-synthesized events.  This results in a perf.data file that is
simpler for external tools to parse.  In particular, this can be used to
prepare a perf.data file for consumption by autofdo.

A subsequent patch makes a change to Intel PT also to enable use with
autofdo and gives an example of that use.

[ Made it use perf_evlist__remove() + perf_evsel__delete() ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>