[tools] perf hists browser: Implement horizontal scrolling
Message-id: <1456127829-13546-161-git-send-email-jolsa@redhat.com>
Patchwork-id: 135388
O-Subject: [PATCH BZ1308570 160/260] perf hists browser: Implement horizontal scrolling
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit c6c3c02dea4034431110923ffd8e296ebfbdbe1b
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Aug 11 17:22:43 2015 -0300

description
===========
Do it using the recently introduced ui_brower scrolling mode, setting
ui_browser.columns to the number of sort columns and then, when
rendering each line, skipping as many initial columns as the user
pressed the right arrow.

As the user presses the left arrow, the ui_browser code will remove the
scrolling counter and the left scrolling takes place.

The right arrow key was an alias for ENTER, so people used to press it
may get a bit annoyed at first, sorry! Ditto for ESC and the left key.

Callchains can be left as is or we can, when rendering the Symbol
column, store the at what position on the screen it is and then
using ui_browser__gotorc() to print it from there, i.e. the callchain
would move around with the symbol.

Leaving it as is, i.e. at a fixed position, close to the left, saves
precious screen real state for it, so I'm inclined to leave it as is
now.

Signed-off-by: Rafael Aquini <aquini@redhat.com>