[tools] perf bench mem: Reorganize the code a bit
Message-id: <1456127829-13546-188-git-send-email-jolsa@redhat.com>
Patchwork-id: 135433
O-Subject: [PATCH BZ1308570 187/260] perf bench mem: Reorganize the code a bit
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 5dd93304a5d386c73c0a59117752acdca67f857d
Author: Ingo Molnar <mingo@kernel.org>
Date: Mon Oct 19 10:04:27 2015 +0200

description
===========
Reorder functions a bit, so that we synchronize the layout of the
memcpy() and memset() portions of the code.

This improves the code, especially after we'll add an strlcpy() variant
as well.

Signed-off-by: Rafael Aquini <aquini@redhat.com>