[tools] perf symbols: we can now read separate debug-info files based on a build ID
Message-id: <1456127829-13546-225-git-send-email-jolsa@redhat.com>
Patchwork-id: 135491
O-Subject: [PATCH BZ1308570 224/260] perf symbols: we can now read separate debug-info files based on a build ID
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 5baecbcd9c9a2f491afe1369fc22e7363f9c94d5
Author: Dima Kogan <dima@secretsauce.net>
Date: Mon Sep 7 17:34:19 2015 -0700

description
===========
Recent GDB (at least on a vanilla Debian box) looks for debug information in

  /usr/lib/debug/.build-id/nn/nnnnnnn

where nn/nnnnnn is the build-id of the stripped ELF binary. This is
documented here:

  https://sourceware.org/gdb/onlinedocs/gdb/Separate-Debug-Files.html

This was not working in perf because we didn't read the build id until
AFTER we searched for the separate debug information file. This patch
reads the build ID and THEN does the search.

Signed-off-by: Rafael Aquini <aquini@redhat.com>