[tools] perf unwind: Pass symbol source to libunwind
Message-id: <1456127829-13546-230-git-send-email-jolsa@redhat.com>
Patchwork-id: 135449
O-Subject: [PATCH BZ1308570 229/260] perf unwind: Pass symbol source to libunwind
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 7ed4915ad60788d6b846e2cd034f49ee15698143
Author: Rabin Vincent <rabin.vincent@axis.com>
Date: Thu Oct 29 08:49:36 2015 +0100

description
===========
Even if --symfs is used to point to the debug binaries, we send in the
non-debug filenames to libunwind, which leads to libunwind not finding
the debug frame.  Fix this by preferring the file in --symfs, if it is
available.

Signed-off-by: Rafael Aquini <aquini@redhat.com>