[tools] perf annotate: Inform the user about objdump failures in --stdio
Message-id: <1456127829-13546-241-git-send-email-jolsa@redhat.com>
Patchwork-id: 135458
O-Subject: [PATCH BZ1308570 240/260] perf annotate: Inform the user about objdump failures in --stdio
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 62ec9b3f02a9bccaf699bd4691db98f779c3075f
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Nov 5 19:06:07 2015 -0800

description
===========
When the browser fails to annotate it is difficult for users to find out
what went wrong.

Add some errors for objdump failures that are displayed in the UI.

Note it would be even better to handle these errors smarter, like
falling back to the binary when the debug info is somehow corrupted. But
for now just giving a better error is an improvement.

Committer note:

This works for --stdio, where errors just scroll by the screen:

  # perf annotate --stdio intel_idle
  Failure running objdump  --start-address=0xffffffff81418290 --stop-address=0xffffffff814183ae -l -d --no-show-raw -S -C /root/.debug/.build-id/28/2777c262e6b3c0451375163c9a81c893218ab1 2>/dev/null|grep -v /root/.debug/.build-id/28/2777c262e6b3c0451375163c9a81c893218ab1|expand
   Percent |      Source code & Disassembly of vmlinux for cycles:pp
  ------------------------------------------------------------------

And with that one can use that command line to try to find out more about what
happened instead of getting a blank screen, an improvement.

We need tho to improve this further to get it to work with other UIs, like
--tui and --gtk, where it continues showing a blank screen, no messages, as
the pr_err() used is enough just for --stdio.

Signed-off-by: Rafael Aquini <aquini@redhat.com>