[tools] perf symbols: Rebuild rbtree when adjusting symbols for kcore
Message-id: <1456127829-13546-248-git-send-email-jolsa@redhat.com>
Patchwork-id: 135466
O-Subject: [PATCH BZ1308570 247/260] perf symbols: Rebuild rbtree when adjusting symbols for kcore
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1308570
https://bugzilla.redhat.com/show_bug.cgi?id=1308570

upstream
========
commit 866548dd6e22c3795ae5146a9746a5cf659698f1
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Nov 6 15:59:29 2015 +0200

description
===========
Normally symbols are read from the DSO and adjusted, if need be, so that
the symbol start matches the file offset in the DSO file (we want the
file offset because that is what we know from MMAP events). That is done
by dso__load_sym() which inserts the symbols *after* adjusting them.

In the case of kcore, the symbols have been read from kallsyms and the
symbol start is the memory address. The symbols have to be adjusted to
match the kcore file offsets. dso__split_kallsyms_for_kcore() does that,
but now the adjustment is being done *after* the symbols have been
inserted. It appears dso__split_kallsyms_for_kcore() was assuming that
changing the symbol start would not change the order in the rbtree -
which is, of course, not guaranteed.

Signed-off-by: Rafael Aquini <aquini@redhat.com>