[s390] kernel: inadvertent free of the vector register save area
Message-id: <1456492815-10194-2-git-send-email-brueckner@redhat.com>
Patchwork-id: 136358
O-Subject: [RHEL7.3 PATCH] [s390] kernel: inadvertent free of the vector register save area
Bugzilla: 1310706
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
If user space processes utilize vector registers, random user space
and kernel crashes might occur in out-of-memory situations.

On fork, the child task structure is created from the parents task
structure. The default arch_dup_tak_struct() function makes a 1:1
copy of the parents task structure, this includes the pointer to the
vector save area.  Eventually the copy_thread function resets the
vector save area pointer to NULL.  The fork operation can fail for
a number of reasons.  It can happen that the new task structure for
the child process is freed again before the copy_thread function has
been reached.  In this case the vector save area of the parent is
freed but the memory is still required by the parent.

Provide an s390 specific version of the arch_dup_task_struct function
which resets the vector save area pointer.

Bugzilla
--------
BZ 1310706
https://bugzilla.redhat.com/show_bug.cgi?id=1310706

Upstream status of the patch
----------------------------
This patch is specific to RHEL7.  The upstream version differs in
this regard due to a major FPU load/save register rework which is
not applicable to RHEL7.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=10545829

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>