[fs] nfsv4.1: Ask for no delegation on OPEN if using O_DIRECT
Message-id: <179d4a9d303f54d15b03cd29080d5a802ba35159.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137202
O-Subject: [RHEL7 PATCH fs 054/396] NFSv4.1: Ask for no delegation on OPEN if using O_DIRECT
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

If we're using NFSv4.1, then we have the ability to let the server know
whether or not we believe that returning a delegation as part of our OPEN
request would be useful.
The feature needs to be used with care, since the client sending the request
doesn't necessarily know how other clients are using that file, and how
they may be affected by the delegation.
For this reason, our initial use of the feature will be to let the server
know when the client believes that handing out a delegation would not be
useful.
The first application for this function is when opening the file using
O_DIRECT.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream-Id: 6ae373394c4257bad562817aa60464ff7fe8f9c4
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>