[fs] nfs: Don't require a filehandle to refresh the inode in nfs_prime_dcache()
Message-id: <54c03fd3728aa73372bc2ccd67e19263182fb728.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137256
O-Subject: [RHEL7 PATCH fs 078/396] NFS: Don't require a filehandle to refresh the inode in nfs_prime_dcache()
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

If the server does not return a valid set of attributes that we can
use to either create a file or refresh the inode, then there is no
value in calling nfs_prime_dcache().

However if we're just refreshing the inode using the attributes that
the server returned, then it shouldn't matter whether or not we have
a filehandle, as long as we check the fsid+fileid combination.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream-Id: fa9233699cc1dc236f4cf42245d13e40966938c5
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>